refactor: improve headers organization
