+++ includes/core/config.h
@@ -1,7 +1,7 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
-#include "../httpc.h"
+#include "api/httpc.h"
 
 int httpc_configure(const httpc_config_t* config);
 const httpc_config_t* httpc_get_config(void);
