+++ includes/core/error_handling.h
