+++ includes/core/handlers/methods.h
@@ -1,7 +1,7 @@
 #ifndef METHODS_H
 #define METHODS_H
 
-#include "http.h"
+#include "api/http.h"
 
 char* handle_get_request(const char* buffer);
 char* handle_post_request(const char* buffer);
