+++ srcs/api/response.c
@@ -1,6 +1,6 @@
 #include "../../includes/api/response.h"
 #include "../../includes/core/utils.h"
-#include "../../includes/http.h"
+#include "../../includes/api/http.h"
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
