+++ srcs/core/handlers/handler.c
@@ -1,4 +1,4 @@
-#include "http.h"
+#include "api/http.h"
 #include <arpa/inet.h>
 
 int allocate_buffer(char **buffer, size_t size) {
