+++ srcs/core/handlers/threads/client_theads.c
@@ -1,4 +1,4 @@
-#include "http.h"
+#include "api/http.h"
 
 void create_client_thread(client_info *info) {
     pthread_t thread;
