+++ srcs/core/logger/user_logger.c
@@ -1,4 +1,4 @@
-#include "../../../includes/logger.h"
+#include "../../../includes/core/logger.h"
 #include "../../../includes/structs/enums.h"
 #include "../../../includes/constants/constants.h"
 #include <stdio.h>
