+++ srcs/core/router/router.c
@@ -1,6 +1,6 @@
-#include "../../../includes/http.h"
-#include "../../../includes/router.h"
-#include "../../../includes/httpc.h"
+#include "../../../includes/api/http.h"
+#include "../../../includes/core/router.h"
+#include "../../../includes/api/httpc.h"
 #include "../../../includes/structs/route.h"
 #include <stdbool.h>
 #include <errno.h>
